Adding basic unit test for getting an order.
I usually wouldn't commit a failing unit test, but I wanted to show my initial concept for how to approach this problem.
I would usually leave dependencies like reading the file to the end, as they can be temperamental and I'd rather prove the concept first by working on the meat of the application.