BasicRateCard now supports specific edge functionality.
This is achieved through adding a specific edge entry, passing in a delegate function.
Ideally, RateCardA and RateCardB would be generated by a factory of some sort later on in the code.
Sadly, I'm out of time. :)
Moving forward, I would implement the Xml reader and marry the three classes together.